refactor(auth): move masked-credential restore into encryption utils
restore_masked_credentials joins mask_credential_dict and
reject_masked_credentials in onyx/utils/encryption.py. Drop the
api-level validate_sso_config call (the db layer already validates on
update) and move the config merge inside the try so update errors match
create's OnyxError shape. Test cleanup: shared _stored_config helper,
unused helper param removed.
